Refactor add and remove into processUpdateRequest
- Refactor `addItemToOrder` and `removeItemFromOrder` functions into
single `processUpdateRequest` function which checks the button class
to determine whether it is supposed to add to or subtract from the list
of items
- TO DO: It's not currently possible to delete the last item in the
cart, but it seems like the error (which doesn't show as an error)
might be coming from the back end. Does it assume an empty array for
items means no update is requested?